requirements.txt
# Basic requirements, will be expanded later
# For now, no specific libraries needed for CLI structure
openai
anthropic
google-generativeai
